add create food test
